+++ src/ethernetip/nxethernetip/nxethernetip.cpp
@@ -53,7 +53,7 @@ static void DumpBytes(const uint8_t *data, size_t length)
          textForm[j] = ((b >= ' ') && (b < 127)) ? (TCHAR)b : _T('.');
       }
       textForm[j] = 0;
-      _tprintf(_T("   %04X | %s | %s\n"), i, buffer, textForm);
+      _tprintf(_T("   %04X | %s | %s\n"), static_cast<unsigned int>(i), buffer, textForm);
    }
 }
 
@@ -262,9 +262,9 @@ static bool GetAttribute(const char *symbolicPath)
    request.writeDataAsUInt16(2);    // Item count
    request.advanceWritePosition(4); // Item type 0 followed by length 0 - NULL address
    request.writeDataAsUInt16(0xB2); // Type B2 - UCMM message
-   request.writeDataAsUInt16(path.size + 2);
+   request.writeDataAsUInt16(static_cast<uint16_t>(path.size + 2));
    request.writeDataAsUInt8(CIP_Get_Attribute_Single);
-   request.writeDataAsUInt8(path.size / 2);
+   request.writeDataAsUInt8(static_cast<uint8_t>(path.size / 2));
    request.writeData(path.value, path.size);
    request.completeDataWrite();
 
@@ -363,9 +363,9 @@ static bool GetAllAttributes(const char *symbolicPath)
    request.writeDataAsUInt16(2);    // Item count
    request.advanceWritePosition(4); // Item type 0 followed by length 0 - NULL address
    request.writeDataAsUInt16(0xB2); // Type B2 - UCMM message
-   request.writeDataAsUInt16(path.size + 2);
+   request.writeDataAsUInt16(static_cast<uint16_t>(path.size + 2));
    request.writeDataAsUInt8(CIP_Get_Attributes_All);
-   request.writeDataAsUInt8(path.size / 2);
+   request.writeDataAsUInt8(static_cast<uint8_t>(path.size / 2));
    request.writeData(path.value, path.size);
    request.completeDataWrite();
 
@@ -452,9 +452,9 @@ static bool GetTCPIPObject()
    request.writeDataAsUInt16(2);    // Item count
    request.advanceWritePosition(4); // Item type 0 followed by length 0 - NULL address
    request.writeDataAsUInt16(0xB2); // Type B2 - UCMM message
-   request.writeDataAsUInt16(path.size + 2);
+   request.writeDataAsUInt16(static_cast<uint16_t>(path.size + 2));
    request.writeDataAsUInt8(CIP_Get_Attributes_All);
-   request.writeDataAsUInt8(path.size / 2);
+   request.writeDataAsUInt8(static_cast<uint8_t>(path.size / 2));
    request.writeData(path.value, path.size);
    request.completeDataWrite();
 
@@ -527,9 +527,9 @@ static bool FindClassInstances(const char *symbolicClassId)
    request.writeDataAsUInt16(2);    // Item count
    request.advanceWritePosition(4); // Item type 0 followed by length 0 - NULL address
    request.writeDataAsUInt16(0xB2); // Type B2 - UCMM message
-   request.writeDataAsUInt16(path.size + 3);
+   request.writeDataAsUInt16(static_cast<uint16_t>(path.size + 3));
    request.writeDataAsUInt8(CIP_Find_Next_Object_Instance);
-   request.writeDataAsUInt8(path.size / 2);
+   request.writeDataAsUInt8(static_cast<uint8_t>(path.size / 2));
    request.writeData(path.value, path.size);
    request.writeDataAsUInt8(127);   // Max number of instances
    request.completeDataWrite();
